Update of the function check interferogram
We change the way the data are aligned on the ZPD. Now the image aligned matrix is automaticcaly aligned with the ZPD at 0 mirro position, and the missing data if any are replaced with zeros. It should not change the results compare to the previous versions of the code though.